# src/main/java/com/backbase/moviesdigger/openapi/moviesDigger-client-v1.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Client API for moviesdigger application
  description: This service is a test task for Backbase team and it's role is to serve users to interact with movies
paths:
  /client-api/v1/movies/{movieName}/iswon:
    get:
      tags:
        - movieDiggerClient
      summary: Indicates if movie is best picture oscar winner
      description: Returns the best picture oscar winner movie list by movie name parameter
      operationId: getWinner
      parameters:
        - name: movieName
          schema:
            type: string
          in: path
          description: movie name
          required: true
        - name: year
          description: The release year of the movie
          schema:
            type: integer
            minimum: 1927
            maximum: 2010
          in: query
          required: false
      responses:
        '200':
          description: A list of oscar's best picture winner movies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetMovieWinnersResponseList"
              example:
                - id: 466b8c0e-238e-46f3-9dba-dae01439bc26
                  name: 'The Lord of the Rings: The Return of the King'
                  year: 2007
                - id: e83dc6ae-caec-405f-9226-d7f60906c5ef
                  name: 'The Lord of the Rings: The Two Towers'
                  year: 2005
        '400':
          description: Bad request. The parameters are not valid
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/Error'
        '404':
          description: A movie name is not found
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/Error'
        '500':
          description: An unexpected condition was encountered
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/Error'

  /client-api/v1/movies/top-rated:
    get:
      tags:
        - movieDiggerClient
      summary: Get a list of top-rated movies
      description: Retrieve a list of the top 10 top-rated movies with optional sorting by box office value
      operationId: getTopRatedMovies
      parameters:
        - name: page
          description: Paged response number.
          schema:
            type: integer
            default: 1
            minimum: 1
          in: query
          required: false
        - name: page-size
          description: The size of the paged response in number of top movies returned. If omitted 10 movies are returned.
          schema:
            type: integer
            default: 10
            minimum: 1
          in: query
          required: false
        - name: sort-direction
          description: The direction in which to sort movies, if not provided, sort in ASC direction
          in: query
          required: false
          schema:
            type: string
            default: ASC
            x-extensible-enum:
              - ASC
              - DESC
      responses:
        '200':
          description: Top rated movies response list ordered in asc or desc direction by office box value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTopRatedMoviesResponseBody'
              example:
                - name: Titanic
                  rating: 9.1
                  boxOffice: 2000000090.78
                  year: 1992
        '500':
          description: An unexpected condition was encountered
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/Error'

  /client-api/v1/movies/{movieName}:
    get:
      tags:
        - movieDiggerClient
      summary: Get movies by provided movie name
      description: Retrieve a list of movies by movie name provided my user
      operationId: getMovies
      parameters:
        - name: movieName
          description: The name of the movie
          in: path
          required: true
          schema:
            type: string
        - name: year
          description: The release year of the movie
          schema:
            type: integer
            minimum: 1927
            maximum: 2010
          in: query
          required: false
      responses:
        '200':
          description: Movies response list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMoviesResponseBody'
              example:
                - id: 466b8c0e-238e-46f3-9dba-dae01439bc26
                  name: Titanic
                  rating: 9.1
                  boxOffice: 2000000090.78
                  year: 1992
        '404':
          description: Provided movie not found
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/Error'
        '500':
          description: An unexpected condition was encountered
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/Error'

  /client-api/v1/movies/rating:
    post:
      tags:
        - movieDiggerClient
      summary: Provide movie's rating
      description: Create the rating of provided movie by rating value and movie id
      operationId: provideMovieRating
      requestBody:
        description: Rating value for a certain movie represented by their id value
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieRatingRequestBody"
      responses:
        '200':
          description: A rating for certain movie provided successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieRatingResponseBody'
              example:
                - name: Titanic
                  rating: 9.1
                  year: 1991
        '400':
          description: Invalid rating value
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/Error'
        '404':
          description: Provided movie not found
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/Error'
        '409':
          description: Rating already was provided by this user before
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/Error'
        '500':
          description: An unexpected condition was encountered
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/Error'

  /client-api/v1/movies/rating/{ratingId}:
    delete:
      tags:
        - movieDiggerClient
      summary: Delete certain user's movie rating
      description: Delete provided certain user's rating of certain movie. Note! Available only for admin
      operationId: deleteMovieRating
      parameters:
        - name: ratingId
          description: Rating of movie rated by user
          in: path
          required: true
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Rating was deleted successfully
          content:
            application/json:
              example: {}
        '400':
          description: Invalid rating value
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/Error'
        '404':
          description: Provided rating not found
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/Error'
        '500':
          description: An unexpected condition was encountered
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/Error'

components:
  schemas:
    GetMovieWinnersResponseList:
      type: array
      items:
        $ref: '#/components/schemas/MovieWinnerResponseBodyItem'
    MovieWinnerResponseBodyItem:
      type: object
      required:
        - id
        - name
        - year
      properties:
        id:
          $ref: 'common.yaml#/components/schemas/Uuid'
        name:
          description: Best picture movie name
          type: string
        year:
          description: Movie release year
          type: integer
          format: int32
    GetTopRatedMoviesResponseBody:
      type: array
      items:
        $ref: '#/components/schemas/TopRatedMovieResponseBodyItem'
    GetMoviesResponseBody:
      type: array
      items:
        $ref: '#/components/schemas/MovieResponseBodyItem'
    TopRatedMovieResponseBodyItem:
      type: object
      required:
        - name
        - rating
        - boxOffice
        - year
      properties:
        name:
          description: Top rated movie name
          type: string
        rating:
          description: Top rated movie rating
          type: string
        boxOffice:
          description: Movie's box office value
          type: string
        year:
          description: Movie release year
          type: integer
          format: int32
    MovieResponseBodyItem:
      type: object
      required:
        - id
        - name
        - rating
        - boxOffice
        - year
      properties:
        id:
          $ref: 'common.yaml#/components/schemas/Uuid'
        name:
          description: Top rated movie name
          type: string
        rating:
          description: Top rated movie rating
          type: string
        boxOffice:
          description: Movie's box office value
          type: string
        year:
          description: Movie release year
          type: integer
          format: int32
    MovieRatingResponseBody:
      type: object
      required:
        - name
        - rating
        - year
      properties:
        name:
          description: Rated movie name
          type: string
        rating:
          description: Rated movie rating
          type: string
        year:
          description: Movie release year
          type: integer
          format: int32
    MovieRatingRequestBody:
      type: object
      required:
        - movieId
        - rating
      properties:
        movieId:
          $ref: 'common.yaml#/components/schemas/Uuid'
        rating:
          description: The new rating for the movie
          type: integer
          format: int32
          minimum: 1
          maximum: 10